chore(react-editable-layers): move code into examples/editor, remove module
