+++ examples/editable-layers/editor/package.json
@@ -7,21 +7,38 @@
   },
   "dependencies": {
     "@deck.gl-community/editable-layers": "^9.0.0-alpha.1",
+    "@deck.gl-community/react": "^9.0.0-alpha.1",
     "@deck.gl/core": "^9.0.12",
     "@deck.gl/extensions": "^9.0.12",
     "@deck.gl/geo-layers": "^9.0.12",
     "@deck.gl/layers": "^9.0.12",
     "@deck.gl/mesh-layers": "^9.0.12",
     "@deck.gl/react": "^9.0.12",
+    "@loaders.gl/core": "^4.2.0",
+    "@loaders.gl/wkt": "^4.2.0",
+    "@maphubs/tokml": "^0.6.1",
+    "@tmcw/togeojson": "^3.2.0",
+    "@turf/helpers": "^6.5.0",
+    "@types/downloadjs": "1.4.3",
     "@types/react": "^18.2.74",
     "@types/react-dom": "^18.2.24",
+    "@types/wellknown": "0.5.4",
+    "boxicons": "^2.1.4",
+    "clipboard-copy": "^3.2.0",
+    "downloadjs": "^1.4.7",
     "maplibre-gl": "^4.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-map-gl": "^7.1.7"
+    "react-dropzone": "^10.2.2",
+    "react-map-gl": "^7.1.7",
+    "styled-components": "^4.4.1",
+    "wellknown": "^0.5.0"
   },
   "devDependencies": {
     "typescript": "^5.4.4",
     "vite": "^5.2.8"
+  },
+  "volta": {
+    "extends": "../../../package.json"
   }
 }
