+++ examples/editable-layers/editor/toolbox/export-component.tsx
