+++ examples/editable-layers/editor/toolbox/import-component.tsx
@@ -1,5 +1,7 @@
 /* eslint-env browser */
 
+// @ts-nocheck
+
 import * as React from 'react';
 import Dropzone from 'react-dropzone';
 import styled from 'styled-components';
