+++ modules/editable-layers/test/edit-modes/lib/draw-line-string-mode.spec.ts
@@ -31,7 +31,7 @@ describe('while tentative', () => {
     expect(props.onEdit.mock.calls[2][0].editContext.position).toEqual([5, 6]);
   });
 
-  it("doesn't change the data", () => {
+  it('doesn\'t change the data', () => {
     const expectedData = {
       type: 'FeatureCollection',
       features: []
@@ -118,7 +118,7 @@ describe('after hitting escape', () => {
     expect(props.onEdit.mock.calls[3][0].editType).toEqual('cancelFeature');
   });
 
-  it("doesn't change the data", () => {
+  it('doesn\'t change the data', () => {
     const expectedData = {
       type: 'FeatureCollection',
       features: []
