--- modules/react-editable-layers/README.md
@@ -1,3 +1,0 @@
-# @deck.gl-community/editable-layers
-
-Provides editable and interactive map overlay layers, built using the power of [deck.gl](https://deck.gl/).
