--- modules/react-editable-layers/package.json
@@ -1,60 +1,0 @@
-{
-  "private": true,
-  "name": "@deck.gl-community/react-editable-layers",
-  "description": "A suite of 3D-enabled data editing layers, suitable for deck.gl",
-  "license": "MIT",
-  "version": "0.0.1",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/visgl/deck.gl-community"
-  },
-  "scripts": {
-    "test": "vitest run",
-    "test-watch": "vitest"
-  },
-  "keywords": [
-    "webgl",
-    "visualization",
-    "overlay",
-    "layer"
-  ],
-  "type": "module",
-  "types": "dist/index.d.ts",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "src"
-  ],
-  "dependencies": {
-    "@deck.gl-community/editable-layers": "^9.0.0-alpha.1",
-    "@deck.gl-community/react": "0.0.1",
-    "@loaders.gl/core": "^4.2.0",
-    "@loaders.gl/wkt": "^4.2.0",
-    "@maphubs/tokml": "^0.6.1",
-    "@tmcw/togeojson": "^3.2.0",
-    "@turf/helpers": "^6.5.0",
-    "@types/downloadjs": "1.4.3",
-    "@types/wellknown": "0.5.4",
-    "boxicons": "^2.1.4",
-    "clipboard-copy": "^3.2.0",
-    "downloadjs": "^1.4.7",
-    "react-dropzone": "^10.2.2",
-    "styled-components": "^4.4.1",
-    "wellknown": "^0.5.0"
-  },
-  "devDependencies": {
-    "react": "^18.2.0"
-  },
-  "peerDependencies": {
-    "react": ">=17.0.2"
-  },
-  "gitHead": "8374ab0ac62a52ae8a6b14276694cabced43de35"
-}
