--- modules/react-editable-layers/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "src",
-    "outDir": "dist",
-    "noEmit": false
-  },
-  "references": [
-    {"path": "../editable-layers"},
-    {"path": "../react"}
-  ]
-}
