+++ modules/react/package.json
@@ -3,7 +3,7 @@
   "name": "@deck.gl-community/react",
   "description": "React components for deck.gl",
   "license": "MIT",
-  "version": "0.0.1",
+  "version": "9.0.0-alpha.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/uber/@deck.gl-community"
