+++ modules/react/src/components/view-control.tsx
@@ -1,4 +1,4 @@
-// @ts-nocheck
+// @ts-nocheck TODO
 
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
