+++ modules/react/src/index.ts
@@ -4,13 +4,13 @@ export {LongPressButton} from './components/long-press-button';
 export {ViewControl} from './components/view-control';
 export {PositionedViewControl} from './components/positioned-view-control';
 
-// Components (originally from @deck.gl-community/react-editable-layers)
+// Components (originally from @nebula.gl/editor)
 
 export {EditorModal as Modal} from './components/modal';
 export {Button} from './components/modal';
 export {Icon} from './components/icon';
 
-// Overlays (originally from @deck.gl-community/react-editable-layers)
+// Overlays (originally from @nebula.gl/overlays)
 
 export {HtmlOverlay} from './overlays/html-overlay';
 export {HtmlOverlayItem} from './overlays/html-overlay-item';
