+++ package.json
@@ -12,6 +12,7 @@
   },
   "type": "module",
   "workspaces": [
+    "examples/editable-layers/editor",
     "modules/*"
   ],
   "scripts": {
